refactor: move reward classifier to sac module and update imports
