+++ lerobot/common/policies/factory.py
@@ -27,8 +27,8 @@ from lerobot.common.policies.diffusion.configuration_diffusion import DiffusionC
 from lerobot.common.policies.pi0.configuration_pi0 import PI0Config
 from lerobot.common.policies.pi0fast.configuration_pi0fast import PI0FASTConfig
 from lerobot.common.policies.pretrained import PreTrainedPolicy
-from lerobot.common.policies.reward_model.configuration_classifier import RewardClassifierConfig
 from lerobot.common.policies.sac.configuration_sac import SACConfig
+from lerobot.common.policies.sac.reward_model.configuration_classifier import RewardClassifierConfig
 from lerobot.common.policies.smolvla.configuration_smolvla import SmolVLAConfig
 from lerobot.common.policies.tdmpc.configuration_tdmpc import TDMPCConfig
 from lerobot.common.policies.vqbet.configuration_vqbet import VQBeTConfig
@@ -67,7 +67,7 @@ def get_policy_class(name: str) -> PreTrainedPolicy:
 
         return SACPolicy
     elif name == "reward_classifier":
-        from lerobot.common.policies.reward_model.modeling_classifier import Classifier
+        from lerobot.common.policies.sac.reward_model.modeling_classifier import Classifier
 
         return Classifier
     elif name == "smolvla":
