+++ lerobot/common/policies/sac/reward_model/modeling_classifier.py
@@ -22,7 +22,7 @@ from torch import Tensor, nn
 from lerobot.common.constants import OBS_IMAGE
 from lerobot.common.policies.normalize import Normalize, Unnormalize
 from lerobot.common.policies.pretrained import PreTrainedPolicy
-from lerobot.common.policies.reward_model.configuration_classifier import RewardClassifierConfig
+from lerobot.common.policies.sac.reward_model.configuration_classifier import RewardClassifierConfig
 
 
 class ClassifierOutput:
