+++ tests/policies/hilserl/test_modeling_classifier.py
@@ -16,8 +16,8 @@
 
 import torch
 
-from lerobot.common.policies.reward_model.configuration_classifier import RewardClassifierConfig
-from lerobot.common.policies.reward_model.modeling_classifier import ClassifierOutput
+from lerobot.common.policies.sac.reward_model.configuration_classifier import RewardClassifierConfig
+from lerobot.common.policies.sac.reward_model.modeling_classifier import ClassifierOutput
 from lerobot.configs.types import FeatureType, NormalizationMode, PolicyFeature
 from tests.utils import require_package
 
@@ -37,7 +37,7 @@ def test_classifier_output():
 
 @require_package("transformers")
 def test_binary_classifier_with_default_params():
-    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
+    from lerobot.common.policies.sac.reward_model.modeling_classifier import Classifier
 
     config = RewardClassifierConfig()
     config.input_features = {
@@ -78,7 +78,7 @@ def test_binary_classifier_with_default_params():
 
 @require_package("transformers")
 def test_multiclass_classifier():
-    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
+    from lerobot.common.policies.sac.reward_model.modeling_classifier import Classifier
 
     num_classes = 5
     config = RewardClassifierConfig()
@@ -117,7 +117,7 @@ def test_multiclass_classifier():
 
 @require_package("transformers")
 def test_default_device():
-    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
+    from lerobot.common.policies.sac.reward_model.modeling_classifier import Classifier
 
     config = RewardClassifierConfig()
     assert config.device == "cpu"
@@ -129,7 +129,7 @@ def test_default_device():
 
 @require_package("transformers")
 def test_explicit_device_setup():
-    from lerobot.common.policies.reward_model.modeling_classifier import Classifier
+    from lerobot.common.policies.sac.reward_model.modeling_classifier import Classifier
 
     config = RewardClassifierConfig(device="cpu")
     assert config.device == "cpu"
